requirements.txt
filetype==1.0.1
pycrypto==2.6.1